WebDAV: fix installation of syncevo-webdav-lookup
The previous change to the syncevo-webdav-lookup broke out-of-tree
compilation because it created a symlink to syncevo-webdav-lookup.sh
inside the current directory, which is not where the file resides when
compiling out-of-tree.

Instead of trying to be smart with symlinks (was only done for to avoid
a small file copy), simply copy the file.