EDS: fix calendar performance improvements
When creating a parent event after its children, the children
really need to be removed first, even with EDS >= 3.6, otherwise
e_cal_client_create_object_sync() fails with "object exists".

This case is rare, so to avoid the performance overhead when
adding in the normal order or just a parent without children,
we optimistically try the operation and only remove if it fails.

That way the normal case is faster while the exceptional case
is slightly slower (two create attempts instead of just one).