README.rst: lanugage tweaks
Be more clear about the local side of a sync.

Avoid "access a peer" when talking about a sync config, because
"access" implies that sync is initiated using the sync config, which
is not the case for SyncML server sync configs in a client-initiated
sync. "talk with peer" is more neutral.