README.rst: data source -> datastore
The word "source" implies reading, while in fact it is read/write.

"datastore" avoids that misconception.

"data store" in two words could also have been used; not sure which
spelling is better.

While renaming, also remove references to explicit --*-property
parameters. The only necessary use today is "--sync-property ?"
and "--datastore-property ?".

--source-property must remain valid for backward compatility.

--datastore-property was used instead of the short --store-property
because "store" might be mistaken for the verb. It doesn't matter
that it is longer because it doesn't get typed often.