EDS calendar: cache by TZID
The display name is not necessarily a unique identifier for the content of a
time zone. For example, the 1990 version of CET/CEST might be different from
the 2014 version. Therefore the display name should not be used as identifier.

The TZID is not necessarily better, but at least the code in
e-cal-check-timezones.c tries to verify content equality before reusing an
existing timezone. It also has code which matches time zones based on their
Olson name embedded in the TZID. If that code does not work well enough
(for example, see FDO #81590), then it should be enhanced.