signon: when access token fails for the first time, explicitly request refresh
This is relevant for the rare situation where signon gave us an access token
that it considered up-to-date and the server rejected it. Without the flag,
signon might just give us the same token again.

One (rare) situation where this might (?) happen is when the clocks
and client and server are out of sync: then the client might think the
token is still valid while the server thinks that it has expired.